Single Ray added
A possibility to add a single ray to the system was added. This is temporary. May be removed latter.
A new symbolic link was added to allow the old designs to work, designs made before changing the workbench structure to the "name space workbench" style. See https://github.com/FreeCAD/freecad.workbench_starterkit